refactor: move getHandler and postHandler to the whatsapp package
