+++ packages/whatsapp/src/index.ts
@@ -1,1 +1,2 @@
+export * from './cf-workers';
 export * from './whatsapp-api-js';
